Added some basic error handling.

+++ src/ConsoleApp/Program.cs
@@ -24,22 +24,39 @@ namespace ConsoleApp
                 CategoryName = "Test Category 2",
                 Description = "This is yet another test category"
             };
-            int id = uow.CategoryRepository.Add(newCategory);
-            Console.WriteLine("Added new category with id of {0}.", id);
+            int idToDelete = uow.CategoryRepository.Add(newCategory);
+            Console.WriteLine("Added new category with id of {0}.", idToDelete);
 
             // Test Update()
-            category = uow.CategoryRepository.GetById(9);
+            int id = 9;
+            category = uow.CategoryRepository.GetById(id);
             category.Description = category.Description + " -- edited";
             Console.WriteLine("Updated existing category. Rows modified: {0}.", uow.CategoryRepository.Update(category));
 
             // Test Delete<T>()
-            category = uow.CategoryRepository.GetById(12);
-            var ret = uow.CategoryRepository.Delete(category);
-            Console.WriteLine("{0} category was deleted using an object reference. There are now {1} categories.", ret, uow.CategoryRepository.GetAll().Count);
+            int ret;
+            category = uow.CategoryRepository.GetById(idToDelete);
+            if (category == null)
+            {
+                Console.WriteLine("Category {0} doesn't exist.", idToDelete);
+            }
+            else
+            {
+                ret = uow.CategoryRepository.Delete(category);
+                Console.WriteLine("{0} category was deleted using an object reference. There are now {1} categories.", ret, uow.CategoryRepository.GetAll().Count);
+            }
 
             // Test Delete(int)
-            ret = uow.CategoryRepository.Delete(11);
-            Console.WriteLine("{0} category was deleted using an id value. There are now {1} categories.", ret, uow.CategoryRepository.GetAll().Count);
+            id = 11;
+            ret = uow.CategoryRepository.Delete(id);
+            if (ret == 0)
+            {
+                Console.WriteLine("Category {0} doesn't exist.", id);
+            }
+            else
+            {
+                Console.WriteLine("{0} category was deleted using an id value. There are now {1} categories.", ret, uow.CategoryRepository.GetAll().Count);
+            }            
 
             Util.WaitForEscape();
         }
